Experimental change

+++ scripts/wsrep_sst_mariabackup.sh
@@ -37,7 +37,7 @@ tcert=""
 tcap=""
 tpem=""
 tkey=""
-tmode="DISABLED"
+tmode='DISABLED'
 sockopt=""
 progress=""
 ttime=0
@@ -82,13 +82,13 @@ backup_threads=""
 encrypt_threads=""
 encrypt_chunk=""
 
-readonly SECRET_TAG="secret"
+readonly SECRET_TAG='secret'
 
 # Required for backup locks
 # For backup locks it is 1 sent by joiner
 sst_ver=1
 
-if [ -n "$(commandex pv)" ] && pv --help | grep -qw -- '-F'; then
+if [ -n "$(commandex pv)" ] && pv --help | grep -qw -F -- '-F'; then
     pvopts="$pvopts $pvformat"
 fi
 pcmd="pv $pvopts"
@@ -101,16 +101,13 @@ if [ -z "$BACKUP_BIN" ]; then
 fi
 
 DATA="$WSREP_SST_OPT_DATA"
-INFO_FILE="xtrabackup_galera_info"
-IST_FILE="xtrabackup_ist"
+INFO_FILE='xtrabackup_galera_info'
+IST_FILE='xtrabackup_ist'
 MAGIC_FILE="$DATA/$INFO_FILE"
 
 INNOAPPLYLOG="$DATA/mariabackup.prepare.log"
 INNOMOVELOG="$DATA/mariabackup.move.log"
 INNOBACKUPLOG="$DATA/mariabackup.backup.log"
-
-# Setting the path for ss and ip
-export PATH="/usr/sbin:/sbin:$PATH"
 
 timeit()
 {
@@ -181,11 +178,11 @@ get_keys()
             exit 2
         fi
         ecmd="'$OPENSSL_BINARY' enc -$ealgo"
-        if "$OPENSSL_BINARY" enc -help 2>&1 | grep -qw -- '-pbkdf2'; then
+        if "$OPENSSL_BINARY" enc -help 2>&1 | grep -qw -F -- '-pbkdf2'; then
             ecmd="$ecmd -pbkdf2"
-        elif "$OPENSSL_BINARY" enc -help 2>&1 | grep -qw -- '-iter'; then
+        elif "$OPENSSL_BINARY" enc -help 2>&1 | grep -qw -F -- '-iter'; then
             ecmd="$ecmd -iter 1"
-        elif "$OPENSSL_BINARY" enc -help 2>&1 | grep -qw -- '-md'; then
+        elif "$OPENSSL_BINARY" enc -help 2>&1 | grep -qw -F -- '-md'; then
             ecmd="$ecmd -md sha256"
         fi
         if [ -z "$ekey" ]; then
@@ -229,12 +226,12 @@ get_transfer()
     if [ $tfmt = 'nc' ]; then
         wsrep_log_info "Using netcat as streamer"
         wsrep_check_programs nc
-        tcmd="nc"
+        tcmd='nc'
         if [ "$WSREP_SST_OPT_ROLE" = 'joiner' ]; then
-            if nc -h 2>&1 | grep -q 'ncat'; then
+            if nc -h 2>&1 | grep -q -F 'ncat'; then
                 wsrep_log_info "Using Ncat as streamer"
                 tcmd="$tcmd -l"
-            elif nc -h 2>&1 | grep -qw -- '-d'; then
+            elif nc -h 2>&1 | grep -qw -F -- '-d'; then
                 wsrep_log_info "Using Debian netcat as streamer"
                 tcmd="$tcmd -dl"
                 if [ $WSREP_SST_OPT_HOST_IPv6 -eq 1 ]; then
@@ -256,14 +253,14 @@ get_transfer()
             # transfer and cause the command to timeout.
             # Older versions of netcat did not need this flag and will
             # return an error if the flag is used.
-            if nc -h 2>&1 | grep -qw -- '-N'; then
+            if nc -h 2>&1 | grep -qw -F -- '-N'; then
                 tcmd="$tcmd -N"
                 wsrep_log_info "Using nc -N"
             fi
             # netcat doesn't understand [] around IPv6 address
-            if nc -h 2>&1 | grep -q ncat; then
+            if nc -h 2>&1 | grep -q -F 'ncat'; then
                 wsrep_log_info "Using Ncat as streamer"
-            elif nc -h 2>&1 | grep -qw -- '-d'; then
+            elif nc -h 2>&1 | grep -qw -F -- '-d'; then
                 wsrep_log_info "Using Debian netcat as streamer"
             else
                 wsrep_log_info "Using traditional netcat as streamer"
@@ -453,7 +450,7 @@ adjust_progress()
         fi
     elif [ -z "$progress" -a -n "$rlimit" ]; then
             # When rlimit is non-zero
-            pcmd="pv -q"
+            pcmd='pv -q'
     fi
 
     if [ -n "$rlimit" -a "$WSREP_SST_OPT_ROLE" = 'donor' ]; then
@@ -580,8 +577,14 @@ get_stream()
 
 sig_joiner_cleanup()
 {
+    local estatus=$?
+    if [ $estatus -ne 0 ]; then
+        wsrep_log_error "Cleanup after exit with status: $estatus"
+    fi
     wsrep_log_error "Removing $MAGIC_FILE file due to signal"
+    [ "$(pwd)" != "$OLD_PWD" ] && cd "$OLD_PWD"
     [ -f "$MAGIC_FILE" ] && rm -f "$MAGIC_FILE"
+    exit $estatus
 }
 
 cleanup_at_exit()
@@ -591,6 +594,8 @@ cleanup_at_exit()
     if [ $estatus -ne 0 ]; then
         wsrep_log_error "Cleanup after exit with status: $estatus"
     fi
+
+    [ "$(pwd)" != "$OLD_PWD" ] && cd "$OLD_PWD"
 
     if [ "$WSREP_SST_OPT_ROLE" = 'joiner' ]; then
         wsrep_log_info "Removing the sst_in_progress file"
@@ -621,7 +626,7 @@ cleanup_at_exit()
     fi
 
     # Final cleanup
-    pgid=$(ps -o pgid= $$ 2>/dev/null | grep -o '[0-9]*' || :)
+    pgid=$(ps -o pgid= $$ 2>/dev/null | grep -o -E '[0-9]*' || :)
 
     # This means no setsid done in mysqld.
     # We don't want to kill mysqld here otherwise.
@@ -712,7 +717,7 @@ recv_joiner()
     local ltcmd="$tcmd"
     if [ $tmt -gt 0 ]; then
         if [ -n "$(commandex timeout)" ]; then
-            if timeout --help | grep -qw -- '-k'; then
+            if timeout --help | grep -qw -F -- '-k'; then
                 ltcmd="timeout -k $(( tmt+10 )) $tmt $tcmd"
             else
                 ltcmd="timeout -s9 $tmt $tcmd"
@@ -812,7 +817,9 @@ monitor_process()
 
 [ -f "$MAGIC_FILE" ] && rm -f "$MAGIC_FILE"
 
-if [ "$WSREP_SST_OPT_ROLE" != 'joiner' -a "$WSREP_SST_OPT_ROLE" != 'donor' ]; then
+if [ "$WSREP_SST_OPT_ROLE" != 'joiner' -a \
+     "$WSREP_SST_OPT_ROLE" != 'donor' ]
+then
     wsrep_log_error "Invalid role '$WSREP_SST_OPT_ROLE'"
     exit 22
 fi
@@ -820,25 +827,17 @@ fi
 read_cnf
 setup_ports
 
-if "$BACKUP_BIN" --help 2>/dev/null | grep -qw -- '--version-check'; then
+if "$BACKUP_BIN" --help 2>/dev/null | grep -qw -F -- '--version-check'; then
     disver=' --no-version-check'
-fi
-
-# if no command line argument and INNODB_DATA_HOME_DIR environment variable
-# is not set, try to get it from my.cnf:
-if [ -z "$INNODB_DATA_HOME_DIR" ]; then
-    INNODB_DATA_HOME_DIR=$(parse_cnf '--mysqld' 'innodb-data-home-dir')
 fi
 
 OLD_PWD="$(pwd)"
 
-cd "$WSREP_SST_OPT_DATA"
-if [ -n "$INNODB_DATA_HOME_DIR" ]; then
-    # handle both relative and absolute paths
-    [ ! -d "$INNODB_DATA_HOME_DIR" ] && mkdir -p "$INNODB_DATA_HOME_DIR"
-    cd "$INNODB_DATA_HOME_DIR"
+if [ -n "$DATA" -a "$DATA" != '.' ]; then
+    [ ! -d "$DATA" ] && mkdir -p "$DATA"
+    cd "$DATA"
 fi
-INNODB_DATA_HOME_DIR=$(pwd -P)
+DATA_DIR="$(pwd -P)"
 
 cd "$OLD_PWD"
 
@@ -866,7 +865,7 @@ if [ $ssyslog -eq 1 ]; then
 else
     if [ $sstlogarchive -eq 1 ]
     then
-        ARCHIVETIMESTAMP=$(date "+%Y.%m.%d-%H.%M.%S.%N")
+        ARCHIVETIMESTAMP=$(date '+%Y.%m.%d-%H.%M.%S.%N')
 
         if [ -n "$sstlogarchivedir" ]; then
             if [ ! -d "$sstlogarchivedir" ]; then
@@ -1100,22 +1099,53 @@ then
         wsrep_log_info "Stale sst_in_progress file: $SST_PROGRESS_FILE"
     [ -n "$SST_PROGRESS_FILE" ] && touch "$SST_PROGRESS_FILE"
 
-    ib_home_dir="$INNODB_DATA_HOME_DIR"
+    # if no command line argument and INNODB_DATA_HOME_DIR environment
+    # variable is not set, try to get it from the my.cnf:
+    if [ -z "$INNODB_DATA_HOME_DIR" ]; then
+        INNODB_DATA_HOME_DIR=$(parse_cnf '--mysqld' 'innodb-data-home-dir')
+        INNODB_DATA_HOME_DIR=$(trim_dir "$INNODB_DATA_HOME_DIR")
+    fi
+
+    if [ -n "$INNODB_DATA_HOME_DIR" -a "$INNODB_DATA_HOME_DIR" != '.' ]; then
+        # handle both relative and absolute paths:
+        cd "$DATA"
+        [ ! -d "$INNODB_DATA_HOME_DIR" ] && mkdir -p "$INNODB_DATA_HOME_DIR"
+        cd "$INNODB_DATA_HOME_DIR"
+        ib_home_dir="$(pwd -P)"
+        cd "$OLD_PWD"
+    fi
 
     # if no command line argument and INNODB_LOG_GROUP_HOME is not set,
-    # try to get it from my.cnf:
+    # then try to get it from the my.cnf:
     if [ -z "$INNODB_LOG_GROUP_HOME" ]; then
         INNODB_LOG_GROUP_HOME=$(parse_cnf '--mysqld' 'innodb-log-group-home-dir')
+        INNODB_LOG_GROUP_HOME=$(trim_dir "$INNODB_LOG_GROUP_HOME")
     fi
 
-    ib_log_dir="$INNODB_LOG_GROUP_HOME"
+    if [ -n "$INNODB_LOG_GROUP_HOME" -a "$INNODB_LOG_GROUP_HOME" != '.' ]; then
+        # handle both relative and absolute paths:
+        cd "$DATA"
+        [ ! -d "$INNODB_LOG_GROUP_HOME" ] && mkdir -p "$INNODB_LOG_GROUP_HOME"
+        cd "$INNODB_LOG_GROUP_HOME"
+        ib_log_dir="$(pwd -P)"
+        cd "$OLD_PWD"
+    fi
 
-    # if no command line argument then try to get it from my.cnf:
+    # if no command line argument and INNODB_UNDO_DIR is not set,
+    # then try to get it from the my.cnf:
     if [ -z "$INNODB_UNDO_DIR" ]; then
         INNODB_UNDO_DIR=$(parse_cnf '--mysqld' 'innodb-undo-directory')
+        INNODB_UNDO_DIR=$(trim_dir "$INNODB_UNDO_DIR")
     fi
 
-    ib_undo_dir="$INNODB_UNDO_DIR"
+    if [ -n "$INNODB_UNDO_DIR" -a "$INNODB_UNDO_DIR" != '.' ]; then
+        # handle both relative and absolute paths:
+        cd "$DATA"
+        [ ! -d "$INNODB_UNDO_DIR" ] && mkdir -p "$INNODB_UNDO_DIR"
+        cd "$INNODB_UNDO_DIR"
+        ib_undo_dir="$(pwd -P)"
+        cd "$OLD_PWD"
+    fi
 
     if [ -n "$backup_threads" ]; then
         impts="--parallel=$backup_threads${impts:+ }$impts"
@@ -1139,7 +1169,7 @@ then
 
     stagemsg='Joiner-Recv'
 
-    MODULE="xtrabackup_sst"
+    MODULE='xtrabackup_sst'
 
     [ -f "$DATA/$IST_FILE" ] && rm -f "$DATA/$IST_FILE"
 
@@ -1162,7 +1192,7 @@ then
                 exit 42
             fi
             CN=$("$OPENSSL_BINARY" x509 -noout -subject -in "$tpem" | \
-                 tr "," "\n" | grep -F 'CN =' | cut -d= -f2 | sed s/^\ // | \
+                 tr ',' '\n' | grep -F 'CN =' | cut -d= -f2 | sed s/^\ // | \
                  sed s/\ %//)
         fi
         MY_SECRET="$(wsrep_gen_secret)"
@@ -1236,14 +1266,29 @@ then
 
         if [ -n "$WSREP_SST_OPT_BINLOG" ]; then
             binlog_dir=$(dirname "$WSREP_SST_OPT_BINLOG")
-            if [ -d "$binlog_dir" ]; then
-                cd "$binlog_dir"
-                wsrep_log_info "Cleaning the binlog directory $binlog_dir as well"
-                rm -fv "$WSREP_SST_OPT_BINLOG".[0-9]* 1>&2 \+ || :
-                [ -f "$WSREP_SST_OPT_BINLOG_INDEX" ] && \
-                    rm -fv "$WSREP_SST_OPT_BINLOG_INDEX" 1>&2 \+
-                cd "$OLD_PWD"
+            binlog_base=$(basename "$WSREP_SST_OPT_BINLOG")
+            binlog_index="$WSREP_SST_OPT_BINLOG_INDEX"
+            binlog_explicit=0
+            if [ -n "$binlog_dir" -a "$binlog_dir" != '.' ]; then
+                binlog_explicit=1
+                if [ -d "$binlog_dir" ]; then
+                    cd "$binlog_dir"
+                    wsrep_log_info \
+                       "Cleaning the binlog directory $binlog_dir as well"
+                    if [ -f "$binlog_index" ]; then
+                        binlogs=$(cat "$binlog_index")
+                        rm -fv "$binlog_index" 1>&2
+                        cd "$DATA_DIR"
+                        echo "$binlogs" | \
+                        while read bin_file || [ -n "$bin_file" ]; do
+                            rm -fv "$bin_file" 1>&2 || :
+                        done
+                    else
+                        rm -fv "$binlog_base".[0-9]* 1>&2 || :
+                    fi
+                fi
             fi
+            cd "$OLD_PWD"
         fi
 
         TDATA="$DATA"
@@ -1279,11 +1324,13 @@ then
 
             dcmd="xargs -n 2 qpress -dT$nproc"
 
-            if [ -n "$progress" ] && pv --help | grep -qw -- '--line-mode'; then
+            if [ -n "$progress" ] && \
+               pv --help | grep -qw -F -- '--line-mode'
+            then
                 count=$(find "$DATA" -type f -name '*.qp' | wc -l)
                 count=$(( count*2 ))
                 pvopts="-f -s $count -l -N Decompression"
-                if pv --help | grep -qw -- '-F'; then
+                if pv --help | grep -qw -F -- '-F'; then
                     pvopts="$pvopts -F '%N => Rate:%r Elapsed:%t %e Progress: [%b/$count]'"
                 fi
                 pcmd="pv $pvopts"
@@ -1293,8 +1340,9 @@ then
 
             # Decompress the qpress files
             wsrep_log_info "Decompression with $nproc threads"
-            timeit "Joiner-Decompression" \
-                   "find '$DATA' -type f -name '*.qp' -printf '%p\n%h\n' | $dcmd"
+            timeit 'Joiner-Decompression' \
+                   "find '$DATA' -type f -name '*.qp' -printf '%p\n%h\n' | \
+                   $dcmd"
             extcode=$?
 
             if [ $extcode -eq 0 ]; then
@@ -1312,24 +1360,36 @@ then
         fi
 
         if  [ -n "$WSREP_SST_OPT_BINLOG" ]; then
-
-            BINLOG_DIRNAME=$(dirname "$WSREP_SST_OPT_BINLOG")
-            BINLOG_FILENAME=$(basename "$WSREP_SST_OPT_BINLOG")
-
-            # To avoid comparing data directory and BINLOG_DIRNAME
-            mv "$DATA/$BINLOG_FILENAME".* "$BINLOG_DIRNAME/" 2>/dev/null || :
-
-            cd "$BINLOG_DIRNAME"
-            for bfile in $(ls -1 "$BINLOG_FILENAME".[0-9]*); do
-                echo "$BINLOG_DIRNAME/$bfile" >> "$WSREP_SST_OPT_BINLOG_INDEX"
-            done
+            cd "$DATA"
+            binlogs=""
+            if [ -f 'xtrabackup_binlog_info' ]; then
+                NL=$'\n'
+                while read bin_string || [ -n "$bin_string" ]; do
+                    bin_file=$(echo "$bin_string" | cut -f1)
+                    if [ -f "$bin_file" ]; then
+                        binlogs="$binlogs${binlogs:+$NL}$bin_file"
+                    fi
+                done < 'xtrabackup_binlog_info'
+            else
+                binlogs=$(ls -d -1 "$binlog_base".[0-9]* 2>/dev/null || :)
+            fi
+            if [ -n "$binlogs" ]; then
+                cd "$DATA_DIR"
+                if [ $binlog_explicit -ne 0 ]; then
+                    [ ! -d "$binlog_dir" ] && mkdir -p "$binlog_dir"
+                    cd "$binlog_dir"
+                fi
+                echo "$binlogs" | \
+                while read bin_file || [ -n "$bin_file" ]; do
+                    echo "$binlog_dir${binlog_dir:+/}$bin_file" >> "$binlog_index"
+                done
+            fi
             cd "$OLD_PWD"
-
         fi
 
         wsrep_log_info "Preparing the backup at $DATA"
         setup_commands
-        timeit "mariabackup prepare stage" "$INNOAPPLY"
+        timeit 'mariabackup prepare stage' "$INNOAPPLY"
 
         if [ $? -ne 0 ]; then
             wsrep_log_error "mariabackup apply finished with errors." \
@@ -1340,7 +1400,7 @@ then
         MAGIC_FILE="$TDATA/$INFO_FILE"
 
         wsrep_log_info "Moving the backup to $TDATA"
-        timeit "mariabackup move stage" "$INNOMOVE"
+        timeit 'mariabackup move stage' "$INNOMOVE"
         if [ $? -eq 0 ]; then
             wsrep_log_info "Move successful, removing $DATA"
             rm -rf "$DATA"
